Add an fsa9480 driver
This is an extremely ugly driver as it uses the input/keyboard
device class.  This is because I couldn't find a way of making
the input path actually switch without reading REG_INT1 and the
only way I could I find a way of polling it was via the keyboard
class.

So, it ain't pretty but it works :)