aries: Add SPL build
This needs to be flashed to offset 0x0 of the OneNAND or, if
unbrickable mod is not applied, you can simply erase/corrupt the first
page of the OneNAND and copy this to an SD card with

dd if=spl/aries-spl.bin of=/dev/mmcblk{x} seek=1

The seek=1 is important as it will skip 512 bytes at the start of the
SD for the MBR (GPT is not supported).

Note that heimdall doesn't appear to work to install this bootloader,
possibly due to some signature checks.

This uses the first stage bootloader used in the HummingBird Interceptor
Bootloader which originally came from the Odroid-T u-boot release.